Rename configstore to package-manager
