+++ packages/gatsby-cli/src/init-starter.ts
@@ -11,7 +11,7 @@ import prompts from "prompts"
 import url from "url"
 
 import report from "./reporter"
-import { getPackageManager, promptPackageManager } from "./util/configstore"
+import { getPackageManager, promptPackageManager } from "./util/package-manager"
 import isTTY from "./util/is-tty"
 
 const spawnWithArgs = (
