+++ packages/gatsby-cli/src/util/package-manager.ts
